base,hash: cast to long all 32bit integers before packing them in a ptr

+++ src/base/pdf-hash-helper.c
@@ -232,7 +232,7 @@ pdf_hash_add_i32 (pdf_hash_t        *table,
 {
   return pdf_hash_add (table,
                        key,
-                       (void *)value,
+                       (void *)((long)value),
                        NULL,
                        error);
 }
@@ -241,7 +241,7 @@ pdf_i32_t
 pdf_hash_get_i32 (pdf_hash_t       *table,
                   const pdf_char_t *key)
 {
-  return (pdf_i32_t) pdf_hash_get_value (table, key);
+  return (pdf_i32_t) ((long) pdf_hash_get_value (table, key));
 }
 
 pdf_bool_t
@@ -252,7 +252,7 @@ pdf_hash_add_u32 (pdf_hash_t        *table,
 {
   return pdf_hash_add (table,
                        key,
-                       (void *)value,
+                       (void *) ((unsigned long)value),
                        NULL,
                        error);
 }
@@ -261,7 +261,7 @@ pdf_u32_t
 pdf_hash_get_u32 (pdf_hash_t       *table,
                   const pdf_char_t *key)
 {
-  return (pdf_u32_t) pdf_hash_get_value (table, key);
+  return (pdf_u32_t) ((unsigned long)pdf_hash_get_value (table, key));
 }
 
 /* Hash helpers to add/get sizes */
